backend/consumidor/requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-api-core==2.12.0
google-auth==2.23.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pika==1.3.2
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytest==7.4.2
requests==2.31.0
rsa==4.9
urllib3==2.0.5
